Add error exit to run.sh

+++ run.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
 docker build --build-arg USE_DEV_CERTS=true -t humanreview-db .
+if [ -z "$(grep 'Successfully built' .out )" ]; then echo "Error with docker build"; exit; fi
+
 docker network create -d bridge ais20
 docker rm -f humanreview-db
 docker run \
